refactor: strip report generation, AI, ticketing, and email from scanner (#245)
These responsibilities have been extracted to the reporter and backend services.

Removed:
- Report generation (JSON/HTML/Markdown/PDF) — 11 report generator files
- AI analysis (Claude API triage + executive summary) — 4 files
- Ticketing service (GitHub issue creation) — 2 files
- Email notifier — 1 file
- Nuclei template generator (used Anthropic API) — 1 file
- Report model, migration, factory, and specs
- Gems: grover, ruby-anthropic, mail
- Docker: chromium, nodejs, npm, pandoc, texlive, puppeteer

Kept:
- Scan orchestration + tool runners
- CVE intelligence enrichment
- Findings normalization + deduplication
- JSON export to GCS (ScanResultsExporter)
- BigQuery logging
- Scan callback service
- Slack notifications
- Cost tracking

389 specs pass, 94.96% coverage, 0 rubocop offenses.

Closes #245
