+++ bin/scan
@@ -52,19 +52,6 @@ if scan.findings_dataset.exclude(cve_id: nil).exclude(cve_id: '').count > 0
   CveIntelligenceService.new.enrich_scan(scan)
 end
 
-# AI Analysis (if API key configured)
-if ENV['ANTHROPIC_API_KEY'].present?
-  puts "\n--- AI Analysis ---"
-  AiAnalyzer.new.analyze_scan(scan)
-end
-
-# Create remediation tickets
-if scan.target.ticketing_enabled?
-  puts "\n--- Remediation Tickets ---"
-  created = TicketingService.new(scan).create_tickets
-  puts "  Created #{created} tickets"
-end
-
 # Export versioned JSON to GCS
 puts "\n--- Scan Results Export ---"
 gcs_scan_results_path = ScanResultsExporter.new(scan).export
@@ -78,14 +65,6 @@ if BigQueryLogger.enabled?
   logged = BigQueryLogger.new.log_from_json(scan_results)
   puts "  Logged #{logged} findings to BigQuery (#{ENV.fetch('SCAN_MODE', 'dev')})"
   audit.bq_loaded(scan, rows_logged: logged)
-end
-
-# Generate reports
-puts "\n--- Report Generation ---"
-generator = ReportGenerator.new(scan)
-reports = generator.generate_all
-reports.each do |report|
-  puts "  #{report.format.upcase}: #{report.gcs_path || 'local'} (#{report.status})"
 end
 
 # Log scan costs to BigQuery
