--- spec/factories/reports.rb
@@ -1,7 +1,0 @@
-FactoryBot.define do
-  factory :report do
-    association :scan, strategy: :create
-    format { 'json' }
-    status { 'pending' }
-  end
-end
