+++ spec/services/notification_service_spec.rb
@@ -15,7 +15,6 @@ RSpec.describe NotificationService do
         allow(ENV).to receive(:fetch).and_call_original
         allow(ENV).to receive(:[]).with('SLACK_WEBHOOK_URL').and_return('https://hooks.slack.com/test')
         allow(ENV).to receive(:fetch).with('SLACK_WEBHOOK_URL', nil).and_return('https://hooks.slack.com/test')
-        allow(ENV).to receive(:[]).with('SMTP_HOST').and_return(nil)
       end
 
       it 'sends a webhook POST request' do
@@ -34,7 +33,6 @@ RSpec.describe NotificationService do
         service.notify
 
         expect(stub).to have_been_requested
-        # Verify the request was made with JSON content type
         expect(WebMock).to have_requested(:post, 'https://hooks.slack.com/test')
           .with(headers: { 'Content-Type' => 'application/json' })
       end
@@ -54,49 +52,10 @@ RSpec.describe NotificationService do
       end
     end
 
-    context 'when email is configured' do
-      before do
-        allow(ENV).to receive(:[]).and_call_original
-        allow(ENV).to receive(:fetch).and_call_original
-        allow(ENV).to receive(:[]).with('SLACK_WEBHOOK_URL').and_return(nil)
-        allow(ENV).to receive(:[]).with('SMTP_HOST').and_return('mail.example.com')
-      end
-
-      it 'delegates to EmailNotifier when SMTP is configured' do
-        email_notifier = instance_double(Notifiers::EmailNotifier)
-        allow(Notifiers::EmailNotifier).to receive_messages(new: email_notifier, configured?: true)
-        expect(email_notifier).to receive(:send_notification)
-
-        described_class.new(scan).notify
-      end
-
-      it 'builds Slack-style email html with correct content' do
-        html = service.send(:build_email_html, 'Test App', scan.summary)
-        expect(html).to include('Test App')
-        expect(html).to include('Penetration Test Scan Complete')
-        expect(html).to include('Critical')
-        expect(html).to include('High')
-      end
-
-      it 'configures smtp settings from ENV' do
-        allow(ENV).to receive(:fetch).with('SMTP_HOST', 'mail.authsmtp.com').and_return('mail.example.com')
-        allow(ENV).to receive(:fetch).with('SMTP_PORT', '2525').and_return('587')
-        allow(ENV).to receive(:fetch).with('SMTP_USERNAME', nil).and_return('user')
-        allow(ENV).to receive(:fetch).with('SMTP_PASSWORD', nil).and_return('pass')
-
-        settings = service.send(:smtp_settings)
-        expect(settings[:address]).to eq('mail.example.com')
-        expect(settings[:port]).to eq(587)
-        expect(settings[:user_name]).to eq('user')
-        expect(settings[:password]).to eq('pass')
-      end
-    end
-
-    context 'when neither webhook nor email is configured' do
+    context 'when webhook is not configured' do
       before do
         allow(ENV).to receive(:[]).and_call_original
         allow(ENV).to receive(:[]).with('SLACK_WEBHOOK_URL').and_return(nil)
-        allow(ENV).to receive(:[]).with('SMTP_HOST').and_return(nil)
       end
 
       it 'does nothing without error' do
@@ -110,7 +69,6 @@ RSpec.describe NotificationService do
         allow(ENV).to receive(:fetch).and_call_original
         allow(ENV).to receive(:[]).with('SLACK_WEBHOOK_URL').and_return('https://hooks.slack.com/test')
         allow(ENV).to receive(:fetch).with('SLACK_WEBHOOK_URL', nil).and_return('https://hooks.slack.com/test')
-        allow(ENV).to receive(:[]).with('SMTP_HOST').and_return(nil)
       end
 
       it 'logs the error and does not raise' do
